backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
psycopg2-binary
Flask-Cors
prometheus_flask_exporter